Use Addressable instead of broken stdlib URI escaping
The native Ruby implementation of URI escaping is broken, and has been deprecated for over 10 years. For example, despite the documentations claims that URI::INVALID characters are removed by default, this is not the case, as square brackets are left in place despite being present in that regex.

We cannot use CGI.escape in this case because CGI.escape uses x-www-form-encoded formatting (where spaces are replaced with +) instead of the correct %20 escaping used in URLs.

* Add Addressable as a dependency
* Use Addressable when generating URLs for asset keys