Fix callback lifecycle
I had attempted to [use a before_delete callback](https://github.com/Shopify/shopify/pull/223539\#discussion_r344247198) and have discovered it did not work.

After investigation, I've noticed the callback chain was executed but it did not respect the return value. So it was entirely possible for a before filter like mine to return 'false' but to be sped through to completion as if it _was_ successful...

Before filters are crucial to stop proceeding when any one returns false. The function implementation that depends on it expects a truthy return value - already a problem because only 'nil', 'false' and 0 cause it to not enter the block. (see https://github.com/Shopify/asset_cloud/blob/385ec2634d29ad4cab51de8d0a0c8feef3bfe8bf/lib/asset_cloud/callbacks.rb\#L28).

In this commit, I explicitly return a boolean depending on the return value of the callback chain.