Enhance test coverage
This commit changes nothing functionally. It makes tests sturdier.

I've noticed that calling #delete on a MemoryBucket was testing callbacks on a key that did not exist. This threw me off because the test expected before_ and after_ callbacks to run when the #delete operation failed - since the key was never written.

All I've done in this commit was to assert for the return value of the principal operation in the test - and set up the tests in order for the original test to verify what was intended.